[Storage] Move testing code to tests
